aggregator/requirements.txt
numpy~=1.23.1
python-dateutil~=2.8.2
confluent_kafka~=1.9.2
faust-streaming~=0.8.11